work[API]: продолжаю реализовывать секретный канал связи; изменил поля и свойства некоторых классов; тестриую практичность взаимодействия

+++ TestAPIProject/ExchangeSystem/Requests/Packages/Default/NewMessage.cs
@@ -9,7 +9,7 @@ namespace ExchangeSystem.Requests.Packages.Default
     {
         public NewMessage(IRequestObject requestObject) : base(requestObject)
         {
-            RequestType = (int)RequestTypes.NewMessage;
+            RequestType = RequestTypes.NewMessage;
         }
     }
 }
